+++ NEWS.md
@@ -5,6 +5,8 @@
 - Improved: Works properly if run from a subdirectory of a crate, or if `-d` points to a
   subdirectory of a crate.
 
+- Improved: Various docs.
+
 ## 0.2.9
 
 Released 2022-07-30