+++ NEWS.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Added: `--cargo-arg` allows passing arguments to cargo commands (check, build, and test), for example to set `--release` or `--features`.
+
 - Improved: Works properly if run from a subdirectory of a crate, or if `-d` points to a
   subdirectory of a crate.